feat: permission route (with not found)

+++ src/controllers/permission.controller.ts
@@ -1,6 +1,7 @@
-import Elysia from "elysia";
+import Elysia, { NotFoundError } from "elysia";
 import PermissionService from "../services/permission.service";
 import { CreatePermissionBody, UpdatePermissionBody } from "../dto/permission.dto";
+import { isNumeric } from "../utils/dataValidator";
 
 const permissionService = new PermissionService();
 
@@ -8,13 +9,16 @@ export const permissionController = new Elysia({ prefix: '/workspaces/:id/permis
 
     // GET /workspaces/:id/permissions
     .get('/', async ({ params }) => {
+        if (!isNumeric(params.id)) throw new NotFoundError()
         const permissions = await permissionService.getPermissionsByWorkspaceId(parseInt(params.id));
         return permissions;
     })
 
     // GET /workspaces/:id/permissions/:permissionId
     .get('/:permissionId', async ({ params }) => {
+        if (!isNumeric(params.id) || !isNumeric(params.permissionId)) throw new NotFoundError()
         const permission = await permissionService.getPermissionById(parseInt(params.id), parseInt(params.permissionId));
+        if (!permission) throw new NotFoundError()
         return permission;
     })
 
@@ -28,7 +32,9 @@ export const permissionController = new Elysia({ prefix: '/workspaces/:id/permis
 
     // PUT /workspaces/:id/permissions/:permissionId
     .put('/:permissionId', async ({ params, body }) => {
+        if (!isNumeric(params.id) || !isNumeric(params.permissionId)) throw new NotFoundError()
         const permission = await permissionService.updatePermission(parseInt(params.id), parseInt(params.permissionId), body);
+        if (!permission) throw new NotFoundError()
         return permission;
     }, {
         body: UpdatePermissionBody
@@ -36,6 +42,8 @@ export const permissionController = new Elysia({ prefix: '/workspaces/:id/permis
 
     // DELETE /workspaces/:id/permissions/:permissionsId
     .delete('/:permissionId', async ({ params }) => {
+        if (!isNumeric(params.id) || !isNumeric(params.permissionId)) throw new NotFoundError()
         const isSuccess = await permissionService.deletePermission(parseInt(params.id), parseInt(params.permissionId));
+        if (!isSuccess) throw new NotFoundError()
         return { message: isSuccess ? 'Success' : 'Failed' };
     });
